feat(longtasks): allow callback to add span attributes on collection
This adds a callback parameter to the LongTaskInstrumentation to be
executed when a longtask is observed. The return value of the callback
is an object of span attributes to be attached to the longtask span.
This feature allows sites using client-side navigation to add window
location details to spans and enables hydrating spans with relevant
information on application state.