fix(instrumentation-graphql): remove graphql as a peerDependency
Now that the types are no longer re-exported the dependency on `graphql`
doesn't make much sense.

This change also has the benefit of letting users who use auto
intrumentation and that have strict peerDependency requirements (if
they're using pnpm for example) to use this intrumentation without
needing to install `graphql`.