fix(pg): do not replace argument with plain object
If the object given as a first argument contained a `.callback`
property, the callback would be patch and the object in the
argument array would be replaced with a plain object. If the object
was a class, the prototype would be lost.

Instead, this implementation replaces the `.callback` property on
the existing object, which is otherwise left as-is.

This fixes issue #1431, where `pg-copy-streams` provides a
`Submittable` object with `.callback` and `.submit` methods defined
in the class prototype. Because `.callback` exists, the object in
the argument array is replaced with a plain object, containing the
patched `.callback`, but no `.submit`.

Because `.submit` was no longer present, `pg` would not treat it as a
submittable, causing it to attempt to execute the query by its
`.text` property, and return `undefined` instead of the submittable.