fix(gql): conform GraphQL span name to spec
Rename graphql "execute" spans to "<operation.type> <operation.name>"
to conform to the spec.

In addition to conforming to the spec, this change can provide
additional value for monitoring purposes. In cases where o11y providers
generate metrics from pre-ingested traces (without all the attributes),
the new naming convention makes it easier to group and filter metrics
by operation type and name, since it is now reflected in the span name.
This can be particularly useful for setting up monitors or alerts.

The query steps other than "execute"'s spans are out of scope as they
do not need to be renamed.