test: fix redis-4 TAV tests with Node.js v14
redis@4.6.9 was a bad release that accidentally broken v14 support.
This updates the versions test to skip that release, and also
reduces the number of versions tested from 27 to 8 by selecting
just the current latest minor releases.

Also, drop the pretest config that (a) was never used
because of the wrong indentation and (b) is no longer
valid since the move to npm workspaces.

Refs: #1860