feat(resource-detector-alibaba): add `service.instance.id` to resource detection
`service.instance.id` is a recommended value. This commit adds it to the alibaba resource detector.
It uses the `serial-number` value from the instance metadata, which has the description:
```
The serial number of the instance.
Example: 4acd2b47-b328-4762-852f-998****
```
Which alignes with the definition for `service.instance.id` from [here](https://github.com/open-telemetry/semantic-conventions/pull/312)

Part Of #2065

Signed-off-by: maryliag <marylia.gutierrez@grafana.com>