feat(bullmq): add bullmq (>=2) instrumentation
Add an instrumentation for BullMQ. This commit adds the already existing
instrumentation at appsignal/opentelemetry-instrumentation-bullmq, with
some minor changes to reduce the functionality and maintenance surface.

The initial work for this instrumentation was done by @jenniferplusplus
at jenniferplusplus/opentelemetry-instrumentation-bullmq, of which both
appsignal/opentelemetry-instrumentation-bullmq and the instrumentation
in this commit are forks.

The instrumentation aims to comply with the Messaging Semantic Conventions
(v0.24.0) whenever possible. Due to the latest published version of the
`@opentelemetry/semantic-conventions` package not fully complying with
the semantic conventions, when in conflict, the instrumentation chooses
to follow the implemented shared package instead of the contents of the
document.