fix(instrumentation-redis-4): avoid diag.error spam when configured client URL is the empty string
Issue #2389 showed that this instrumentation would *crash* on a Redis
client configured with `{url: ''}` (an empty string). The crash was
fixed in #2397. This change avoids the once-crashy code, and hence
the diag.error spam, by using the same guard before attempting to parse
the configured client URL that the Redis client itself does, `if
(options?.url)`,

Refs: #2389