test(instrumentation-express): update to use correct span kind enum; also fix test flakiness
This is a follow-up on https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2408#discussion_r1735862915
to update one more usage of `SpanKind.*` to `testUtils.OtlpSpanKind.*`.

While testing this locally I also noticed that the instr-express tests
using runTestFixture are flaky. The issue is that `TestCollector#sortedSpans`
can get the span ordering wrong when spans start in the same millisecond
(the resolution of span.startTimeUnixNano). This happens with Express
middleware spans on a fast machine.  I've updated `sortedSpans` to
fallback to using span.parentSpanId to attempt to get more reliable
sorting.