doc(*): Migrate away from deprecated `addSpanProcessor` in READMEs
This migrates from the deprecated `addSpanProcessor` method to the
constructor option in preparation for the upcoming SDK 2.0 release.

This commit migrates everything in the README.md of each package,
other than that there are no runtime code changes.

Technically this is a user-facing change for users that previously
followed the examples.

Ref #2645