chore(deps): update collector images in examples
This PR updates all the collectors in various examples to the latest current
release of the collector. This was needed because I was trying to run the
graphql example and ran into a problem where the image of the collector was
so old that an arm release of the container wasn't available.

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>