test(instrumentation-redis-4): fix tests with redis@4.4.0 and earlier (#2914)
In redis@4.4.0 and earlier versions of redis v4, client.disconnect()
will throw if the connect failed. That broke the "with empty string for
client URL, ..." test case, at least on macOS:

    npm run test:docker:run
    RUN_REDIS_TESTS=1 npm t

I cannot explain why this is not failing in CI. Perhaps something
platform specific? This is related to socket handling in the redis
client. I believe the relevant change in node-redis was:
https://github.com/redis/node-redis/pull/2295
which was part of `@redis/client@1.3.1` which was included in
`redis@4.4.0`.