chore: rename workspace dirs to packages/${package-base-name} (extreme edition)
This PR is the final take of the plan discussed at #2223.
This implements option 3 of https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2223#issuecomment-2125925648
(i.e. the "extreme" option).

Before this change the packages published from this repo were spread
across these workspace dirs:

    "packages/*",
    "plugins/node/*",
    "plugins/web/*",
    "propagators/*",
    "detectors/node/*",
    "metapackages/*"

with the directory name *usually* matching the npm package name, and
with or without a leading "opentelemetry-". This PR normalizes all dirs
to be under "packages/*" and to have the base dir match the package
name (excluding the npm `@opentelemetry/` org prefix).

Refs: https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2223