chore: lint .js and .mjs files (in addition to the existing .ts file linting)
In https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2567 support
was added to eslint.config.js for linting .js and .mjs files -- before it
only supported .ts files.
Originally this was added to support linting examples/... which includes many
.js files.

This change enables linting of .js and .mjs files in the rest of the repo.

Refs: https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2891