chore(contrib-test-utils)!: update semconv usage to ATTR_ exports
This also drops the `imageTag` option to `assertContainerResource`.
This option is unused by all users of `assertContainerResource` in the
repo. (It dates back to when `@opentelemetry/resources` in the core
repo used contrib-test-utils. It no longer does.)  Dropping this
option allows us to not have to deal with the Semantic Conventions
change of 'container.image.tag' to container.image.tags' (plural)
in semconv v1.22.0.

Refs: #2377