ci(release-please): combine npm publish workflow jobs to avoid file perm issues from zip files used by upload-artifact
With the previous separate of jobs, the built files to publish were shared
across jobs with upload/download-artifact. This uses zip files, which cannot
handle the execute bit on files. This resulted in lerna publish spuriously
erroring out thinking there were uncommited changes to the handful of
files in scripts/ that intentional have the execute file perm set.

Refs: https://github.com/open-telemetry/opentelemetry-js-contrib/pull/3107#issuecomment-3335487278