chore(release-please): drop unneeded package-lock sync as part of the release PR
Now that we are on npm lockfileVersion=3, which doesn't have the 'dependencies'
legacy key, we *should* no longer need this sync.
Note that for the recent release-please PR there *was* some churn in 'peer'
entries in the package-lock file from this step:
https://github.com/open-telemetry/opentelemetry-js-contrib/pull/3135/commits/54c1f5beafcc91e4cab497edfea738e3b25599b9
My educated guess is that this is *not* impactful, not necessary for
the release-please workflow.