feat(instrumentation-openai): add support for instrumenting openai@6
One complication was that in openai@5.12.1 'interface ChatCompletionMessageToolCall'
was changed to be:
    export type ChatCompletionMessageToolCall =
      | ChatCompletionMessageFunctionToolCall
      | ChatCompletionMessageCustomToolCall;
which complicated capturing 'tool_calls' info for the telemetry event.

Note also that the GenAI semconv has since moved on. As of https://github.com/open-telemetry/semantic-conventions/pull/2179
there is no longer a 'tool_calls' property in a captured event. This PR
does *not* get into updating the semconv version used by this instrumentation.

Unrelated, there was also a fix to a typo in the 'compile:with-dependencies' npm script
for this package.