fix(instrumentation-web-exception): properly add bounded error listener
The README.md advises to use the instrumentation like this:

```
registerInstrumentations({
  instrumentations: [
    new ExceptionInstrumentation({ ... }),
  ],
});
```

But this does not work because the function passed to
`addEventListener('error', ...)` is the unbounded method, so accesses to
`this.getConfig()` failed.

I know this sounds surprising, because the `onError` method was properly
reaffected with the bounded method in the constructor. But actually,
`enable()` is called by the base class, during `super()`, which is
called *before the `onError` is bound*!

This wasn't caught in tests because in test, `enable()` is called two
times, once when instantiating the instrumentation, and once explicitely
when calling `instr.enable()`. This second call *do* register the
correctly bounded method.

To work around this, this PR make sure the `onError` method is proprely
bound in `enable()`.